Creation of new branch for refactors
