+++ app/src/main/java/com/khonsu/enroute/AsyncUpdateTask.java
@@ -1,0 +1,4 @@
+package com.khonsu.enroute;
+
+public class AsyncUpdateTask {
+}
