+++ app/src/main/java/com/khonsu/enroute/messaging/MessageGenerator.java
@@ -1,25 +1,23 @@
-package com.khonsu.enroute;
+package com.khonsu.enroute.messaging;
 
-import android.content.Context;
-import android.os.Handler;
-import android.os.Looper;
-import android.widget.Toast;
+import com.khonsu.enroute.GoogleMapsDurationGetter;
+import com.khonsu.enroute.GoogleMapsLinkGenerator;
+import com.khonsu.enroute.UnableToGetEstimatedJourneyTimeException;
 
 public class MessageGenerator {
 	static final String MESSAGE_ESTIMATE_FORMAT = "I should arrive in approximately %s.";
 	static final String CURRENT_LOCATION_ESTIMATE_FORMAT = "My current location is: %s.";
     static final String PLUG=" Update sent by En Route!";
 
-	static final String COULD_NOT_UPDATE_ETA = "Could not obtain ETA. Please check destination value and internet connection.";
 	private final GoogleMapsDurationGetter googleMapsDurationGetter;
 
 	public MessageGenerator(GoogleMapsDurationGetter googleMapsDurationGetter) {
 		this.googleMapsDurationGetter = googleMapsDurationGetter;
 	}
 
-	public String generateMessage(Context context, String currentLatitude, String currentLongitude, String destination, String mode) {
+	public String generateMessage(String currentLatitude, String currentLongitude, String destination, String mode) throws UnableToGetEstimatedJourneyTimeException {
 		StringBuffer message = new StringBuffer();
-		appendEstimateIfPossible(context, currentLatitude, currentLongitude, destination, mode, message);
+		appendEstimate(currentLatitude, currentLongitude, destination, mode, message);
 		appendCurrentLocationLink(currentLatitude, currentLongitude, message);
         message.append(PLUG);
 		return message.toString();
@@ -32,20 +30,7 @@ public class MessageGenerator {
 		message.append(String.format(CURRENT_LOCATION_ESTIMATE_FORMAT, GoogleMapsLinkGenerator.generateLinkForLatLong(currentLatitude, currentLongitude)));
 	}
 
-	private void appendEstimateIfPossible(final Context context, String currentLatitude, String currentLongitude, String destination, String mode, StringBuffer message) {
-		try {
+	private void appendEstimate(String currentLatitude, String currentLongitude, String destination, String mode, StringBuffer message) throws UnableToGetEstimatedJourneyTimeException {
 			message.append(String.format(MESSAGE_ESTIMATE_FORMAT, googleMapsDurationGetter.getEstimatedJourneyTime(currentLatitude, currentLongitude, destination, mode)));
-		} catch (UnableToGetEstimatedJourneyTimeException e) {
-			e.printStackTrace();
-
-			Handler handler = new Handler(Looper.getMainLooper());
-			handler.post(new Runnable() {
-
-				@Override
-				public void run() {
-					Toast.makeText(context, COULD_NOT_UPDATE_ETA, Toast.LENGTH_LONG).show();
-				}
-			});
-		}
 	}
 }
