+++ purejava/src/main/java/com/khonsu/enroute/GoogleMapsDurationGetter.java
@@ -1,5 +1,7 @@
 package com.khonsu.enroute;
 
+import android.util.SparseArray;
+
 import com.google.gson.Gson;
 import com.khonsu.enroute.google.datamodel.maps.NoDurationInResponseException;
 import com.khonsu.enroute.google.datamodel.maps.ResponseBody;
@@ -8,15 +10,22 @@ import java.io.IOException;
 import java.net.URL;
 
 public class GoogleMapsDurationGetter {
+	private final static SparseArray<String> MODE_OF_TRANSPORT_ID_TO_GOOGLE_STRING = new SparseArray<String>(){{
+		put(R.id.mode_of_travel_car, "driving");
+		put(R.id.mode_of_travel_bike, "bicycling");
+		put(R.id.mode_of_travel_walking, "walking");
+	}};
+
 	private UrlAccessor urlAccessor;
 
 	public GoogleMapsDurationGetter(UrlAccessor urlAccessor) {
 		this.urlAccessor = urlAccessor;
 	}
 
-	public String getEstimatedJourneyTime(String currentLat, String currentLong, String destination, String mode) throws UnableToGetEstimatedJourneyTimeException {
+	public String getEstimatedJourneyTime(String currentLat, String currentLong, String destination, int mode) throws UnableToGetEstimatedJourneyTimeException {
 		try {
-			URL url = GoogleMapsLinkGenerator.generateLinkForDirections(currentLat, currentLong, destination, mode);
+			String modeAsGoogleString = MODE_OF_TRANSPORT_ID_TO_GOOGLE_STRING.get(mode);
+			URL url = GoogleMapsLinkGenerator.generateLinkForDirections(currentLat, currentLong, destination, modeAsGoogleString);
 			Gson gson = new Gson();
 			String responseBodyAsJsonString = urlAccessor.getResponse(url);
 			ResponseBody responseBody = gson.fromJson(responseBodyAsJsonString, ResponseBody.class);
@@ -24,7 +33,6 @@ public class GoogleMapsDurationGetter {
 		} catch (NoDurationInResponseException | IOException e) {
 			e.printStackTrace();
 		}
-
 		throw new UnableToGetEstimatedJourneyTimeException();
 	}
 }
