+++ purejava/src/main/java/com/khonsu/enroute/ModeOfTransport.java
@@ -1,0 +1,4 @@
+package com.khonsu.enroute;
+
+public enum ModeOfTransport {
+}
